+++ NEWS.md
@@ -4,6 +4,9 @@
 locally and only via the 
 [pkgdown site](https://scotgovanalysis.github.io/sgplot/) (#66).
 
+* Add `quietly` argument to `use_sgplot()` to allow messages to be suppressed 
+(#67).
+
 * Set minimum version of ggplot (>= 3.5.0) (#62).
 
 * Fix bug where `use_sgplot()` could not handle passing multiple arguments of